vim: Don't unconditionally set textwidth
The 'gitcommit' ftplugin sets textwidth=72 *only if* textwidth has not
already been set (that is, only if textwidth == 0). There may be other
plugins that intelligently set textwidth in this way.

Rather than set textwidth unconditionally, we should set it per filtype
so as not to clobber features of wise plugins.