bashrc: Fix PROMPT_COMMAND for Mac and Gnome
When opening new terminal tabs and windows I want to preserve the pwd.
This patch fixes that behavior on Mac OS and some versions of
gnome-terminal.

Each OS, and different releases of gnome-terminal, define PROMPT_COMMAND
to a different default value. Stop trying to guess what that default
value is. Instead, append my custom PROMPT_COMMAND to the default value.