bashrc: Fix PROMPT_COMMAND for subshells
When a subshell is spawned, parent shell's ~/.bashrc has already set PROMPT_COMMAND.
When the subhsell's ~/.bashrc attempts to again append its custom
PROMPT_COMMAND to the already-defined PROMPT_COMMAND, PROMPT_COMMAND
acquires an invalid value.

To fix it, add an initial canary string to PROMPT_COMMAND so that we
don't set it again in subhsells.